Bug fixes and assembly_instance pseudo-element
Adds a new pseudo-element called assembly_instance.
This only exists as an element on root, and is a service. It will
equal the current Instance object for the running assembly. Importantly,
it also allows using overlays on this service, providing a way for
assemblies to intercept and transform the instance objects given out
to users.