Clarify scope attribute
Previously, scope was doing double duty as the elements in scope
for an element, and also the children of a group. This required
an additional :full_scope attribute to be the actual scope of groups.

Instead, make scope always be the elements in scope and get rid of
full_scope.